Model Viewer [WIP 2]

+++ GCNCommon/src/net/medsouz/gcn/model/bmd/sections/shp1/Primitive.java
@@ -6,6 +6,6 @@ public class Primitive {
 	public PrimitiveType type;
 	public short numVertices;
 	
-	public ArrayList<Short> positionMatrixIndices = new ArrayList<Short>();
-	public ArrayList<Short> positionIndices = new ArrayList<Short>();
+	public ArrayList<Integer> positionMatrixIndices = new ArrayList<Integer>();
+	public ArrayList<Integer> positionIndices = new ArrayList<Integer>();
 }
